初步提交测试版

+++ app/src/main/java/com/whmnrc/flymall/beans/ActivityGoodsListBean.java
@@ -327,9 +327,9 @@ public class ActivityGoodsListBean {
                     private boolean HasSKU;
                     private int Id;
                     private String ImagePath;
-                    private int MarketPrice;
+                    private double MarketPrice;
                     private String MeasureUnit;
-                    private int MinSalePrice;
+                    private double MinSalePrice;
                     private String ProductCode;
                     private String ProductName;
                     private int Quantity;
@@ -454,11 +454,11 @@ public class ActivityGoodsListBean {
                         this.ImagePath = ImagePath;
                     }
 
-                    public int getMarketPrice() {
+                    public double getMarketPrice() {
                         return MarketPrice;
                     }
 
-                    public void setMarketPrice(int MarketPrice) {
+                    public void setMarketPrice(double MarketPrice) {
                         this.MarketPrice = MarketPrice;
                     }
 
@@ -470,11 +470,11 @@ public class ActivityGoodsListBean {
                         this.MeasureUnit = MeasureUnit;
                     }
 
-                    public int getMinSalePrice() {
+                    public double getMinSalePrice() {
                         return MinSalePrice;
                     }
 
-                    public void setMinSalePrice(int MinSalePrice) {
+                    public void setMinSalePrice(double MinSalePrice) {
                         this.MinSalePrice = MinSalePrice;
                     }
